refactor: split tabs into separate components
